match react-docgen's expressionTo module with local external modifications
